Setup form state handling boilerplate logic
